Bug #26881855: GCOL: INCORRECT BEHAVIOR WITH CONSTANT ITEM
Issue:
------
Virtual generated column expression that contain a
comparision operators between certain datatypes can cause
problems.

When a table is opened, the expression in the generated
column is created on the table's memroot. As subsequent
statements access the same table, convert_constant_item()
checks the types and under certain conditions chooses to
re-create one of the objects with the correct type. The new
object will be created on the statement's memroot and will
be freed at the end of the statement. This can cause
problem when subsequent statements access the same table.

Root cause:
-----------
Currently Item_int_with_ref class has no field_type and
defaults to an integer. convert_constant_item() decides to
re-create the objects based on the field_type even when a
re-create will only result in a new Item_int_with_ref
object.

Solution:
---------
Add a field_type to Item_int_with_ref. This will make sure
that convert_constant_item() can distinguish if a new
object is actually neccessary.

This fix also addresses Bug #26881924 and Bug #26881874.

This problem was fixed in trunk by Roy's Bug#25221172.