Bug#26758939 : FAIL TO LOCK THREAD TO CPU - Post Push Fix
Previous patches for Bug#26758939 disabled Lock/UnlockCPU apis in
environment where the glibc version was less than 2.23 to avoid the
effects of a bug. But since this seems too strict, most of the
changes made as a part of Bug#26758939 are reverted in this patch. This
re enables Lock/UnlockCPU apis in all environments regardless of the
glibc version used.

The glibc bug causes the sched_setaffinity() call to fail when an all
ones CPU mask is set. The call fails and the error is set to EINVAL -
invalid arguments. UnlockCPU uses a all ones CPU mask and will fail
similarly if the glibc version is lower than 2.23. This patch instead
of using an all ones CPU mask, uses sysconf to get the actual number of
processors available to the system and then sets the cpu mask
accordingly.