Bug#26877788 SELECT FROM INFORMATION_SCHEMA.FILES RETURNS NO RECORDS WHEN ORDER BY IS USED
During subquery materialization table list is modified
so that empty JOIN_TAB table is inserted in the middle of
table list. It leads to interruption of the loop in
get_schema_tables_result() function and tables which are
after empty JOIN_TAB are not processed.

The fix:
Do not interrupt the loop on empty JOIN_TAB.