Bug #26941370 ST_GEOMFROMGEOJSON ALLOWS OUT-OF-RANGE COORDINATES
Problem: ST_GeomFromGeoJSON didn't check that geographic coordinates
were within the valid range.

Also, it was allowed to add a "crs" member to any level of the GeoJSON
document, and the specified SRID would take effect for the whole
document. The GeoJSON specification mandates that the "crs" member, if
any, is put at the top leve of the document.

Fix: Check that longitude and latitude coordinates are within the ranges
(-180, 180] and [-90, 90], respectively (in degrees -- other limits for
other units). Raise ER_LONGITUDE_OUT_OF_RANGE or
ER_LATITUDE_OUT_OF_RANGE if coordinates are out of range.

The exact limits may deviate slightly from the ideal because of floating
point arithmetics.

Fail with ER_INVALID_GEOJSON_CRS_NOT_TOP_LEVEL if a "crs" member that
specifies a different SRID than the SRID of the top-level object is
found at a lower level of the GeoJSON document.

Change-Id: Ib5a55b781bf0fcd71dc06fe6dbef3a67778b30e8