Bug#26968613 Move pseudo columns ROW_COUNT and COMMIT_COUNT to Dbtup.
In Dbtup there is a working counter of committed rows.

Also in some ndbinfo tables count of committed rows per fragment and
committed operations per fragment was reported and are now taken from
Dbtup instead of Dbacc.

The one used in was updated for each operation committed for a row, and
since the commit signals can arrive out of order transient counter values
could be negative.

For example if an transaction had several interleaved insert and delete
on same row, the commit signals for delete operations could arrive first
to Dbacc and commit signals for insert operations later.

Reviewed by Mikael Ronstrom 2017-10-17.