Bug#26975864: ASSERTION FAILED: DEC <= 6, MY_DATETIME_TO_STR, LEAST, GREATEST, ETC
If one of the temporal arguments to LEAST/GREATEST was of type DATETIME,
but the aggregated data type was something else than VARCHAR or a
temporal type, the LEAST/GREATEST item would not have a temporal data
type while its temporal properties were being set. The temporal
properties for the LEAST/GREATEST item, including its fractional
precision, would therefore not be calculated correctly, causing an
assertion to fail later in the execution.

As a fix, the data type of LEAST/GREATEST is now temporarily set to a
temporal type if one or more, but not all, of its arguments are of temporal type
regardless of the aggregated data type.