BUG#26787288: CHECK_WARNINGS FOR GROUP_REPLICATION.GR_RELAY_LOG_NAME_CHANGE TIMES OUT ON PB2
The test produces a failure when server is starting and the following steps put
it on restart. This situation caused a possibility of group replication execute
an query on the server but don't have any response due server is stopping.

The group replication service sql command wasn't retrieving an error when this
failure happened that caused the server to crash on the end of the test.

We modified the function to return failure when this occurred.

When accessing to SQL result we confirm that it has values.