BUG#26539665: ASSERT FAILURE: !SDI_INSERT_FAILED || TRX_IS_INTERRUPTED(TRX)
Problem: The above problem was found to be caused by parallel inserts
of the schema SDI into the SDI B-tree when creating tables in the same
schema in parallel.

Solution: Don't create and store schema SDIs. The schema SDI only
contains the default character set for new tables in the schema. This
is information is not essential and not needed for import. Not storing
the SDIs avoid the contention which causes the original problem.

This patch covers the server changes.