WL#11179: Add atrt support to set processes' cpu affinity
This feature introduces a new my.cnf configuration group called
'cluster_deployment' and a new option called 'cpubind'.
This configuration group is read only by atrt and group hierarquies are
supported.
The 'cpubind' option can receive a list of cpus (format "CPU0,CPU1,...") and
will make the process locked to the specified cpus. If no option is set, the
default behavior applies, and processes are free to run in any available cpu.
Although this feature was developed mainly to lock mysqld processes to
specific cpus, it works for locking cpu for tests or any ndb process.

Examples of configuration are:

  # Applying cpubind to all processes in all clusters
  [cluster_deployment]
  cpubind = 0

  # Applying cpubind to all mysqld processes in '.2node' cluster
  [cluster_deployment.mysqld.2node]
  cpubind = 0

  # Applying cpubind to ndb_mgmd's processes in '.2node' cluster
  [cluster_deployment.ndb_mgmd.1.2node]
  cpubind = 0

  [cluster_deployment.ndb_mgmd.2.2node]
  cpubind = 3

NOTE: This feature supports Linux only.