Bug#24510948: Post-push fix.
Original patch changed main.schema because some parallel operations
now block on schema MDL, rather than table MDL. This because the
schema MDL taken when a table is loaded into the TDC was made
transactional.

However, when running with --ps-protocol this schema MDL is held only
during prepare, as transactional MDLs are released when rolling back
to savepoint at the end of prepare. During execute the table is still
in the TDC, so no MDL is taken on schema, and the old behavor of
blocking on table MDL is observed.

Introduce a variable for the wait condition and set it according to the
--ps-protocol value.

(cherry picked from commit 2c07bd377e2c29b5c07e222ce611d7fcf272fe74)