Bug#26771717 ASSERT IN TEMPTABLE::STORAGE::CLEAR()
There was an invalid assertion in clear() metho and it was removed.
The assertion checked that when there are multiple pages the number
of elements must be larger than number of elements on a single page.
This was not true when the erase() was called as it may decremented
the number of elements leaving the pages allocated if the elements
were not at the end.

Additionally fixed missing initialization in move constructor.
The pointer members were uninitialized and in some cases (non-zeroed
memory allocated for the object) the asserts in move assignement
operator were raised.

Reviewed by: Pawel Olchawa <pawel.olchawa@oracle.com>