Bug#25190109 AFTER WINDOW FUNCTIONS ARE PUSHED, SIMPLIFY SOME CTE CODE
In the implementation of window functions, we extended HEAP_POSITION
which makes the need for allow_scan_from_position go away
(see change in sql_tmp_table.cc).
Practical consequence: in a recursive CTE with UNION DISTINCT,
and if there are lots of columns in the CTE, so that their total
length was bigger than a few hundred bytes, we were forcing the
CTE's tmp table to be innodb, now we don't anymore, it will be
MEMORY.

Also, window functions were developed to work with both MEMORY
and Temptable as internal engine; but right before the release of WF
there was a MTR test timeout which led us to force Temptable; later
investigation showed that the failure was just a wrongly sized
max_heap_table_size, so there was nothing to worry about,
so we re-enable MEMORY for WFs.