Bug #26975882: CRASH IN STRING::COPY, USUALLY WITH WKB FUNCTIONS / WINDOW FUNC
An error is generated while evaluating the first argument to Item_func_buffer.
Without checking for errors, we proceed further to evaluate the second argument
which is a WINDOWING function. And the error is caught here. Windowing
function returns nullptr. But does not set null_value to true. This results
in server exit later.

Solution:
Check the error early in Item_func_buffer function.