Bug#20964700: RECURSION OF ITEM_FUNC::WALK LEAD TO STACK OVERFLOW
Constant propagation is sometimes too eager and creates cycles in the
Item tree, which again leads to infinite recursion when the Item tree
is processed.

This patch skips constant propagation when the constant expression
contains a reference to the column it is meant to replace, thus
avoiding the creation of cycles in the Item tree.