Fix memory leak for ./mtr --mem --sanitize innodb.autoinc_persist_debug
The problem is PersistentTableMetadata was allocated by replacement
new during recovery, but the vector in it may require more memory
later. So the new memory could not be managed by the delete of
original memory used for replecement new.

Since the replacement new is totally unncessary, replacing it with
normal new fixes this problem.

Reviewed-by: Allen Lai <allen.lai@oracle.com>