BUG#27008102: CRASH DURING SHUTDOWN WHEN START GR EXECUTED IF MAX CONNECTIONS ALREADY OPENED BUG#27016552: MAX_CONNECTIONS ARE ACCOUNTED FOR GR INTERNAL CONNECTIONS
Group Replication does use the SQL API to execute internal
operations on the server during START and STOP, like enable and
disable read only mode.
While opening this internal connection, if the server already
reached the max_connections value, Group Replication was failing as
expected but a dangling thread was left behind, which later would
cause issues.

This patch addresses that issue, by handling completely the open
connection failure scenario.
This patch also improves the error messages on this error scenario,
by including the internal server error and message, and on the
particular case of max_connections also instructing the user what
needs to be changed.