Bug #26740557: WINDOW FUNC + JSON: ASSERTION FAILED: FALSE IN ITEM::VAL_JSON
Problem:
For the testcase in bugpage, after aggregating datatypes in
Item_lead_lag::resolve_type, the the resulting datatype is set to be
ITEM_TYPE_JSON. As a result, we further determine that the compare
function for "=" operator would be compare_json which calls val_json() on
Item_lead_lag. Since val_json is not defined, Item::val_json gets called
which asserts.

Analysis:
In a general case, windowing functions get evaluated in end_write_wf() which
calls save_in_field. This function calls evaluation functions based on the
result type (except for some special cases). So for json datatypes, it would
be val_str. However in the test case, no user tables are involved. So the
plan is determined to be constant and therefore we set the end select function
to be end_send_group which calls the evaluation functions based on the datatypes
involved.

Solution:
Defined val_json for Item_lead_lag. Also changed ::resolve_type() for
Item_first_last_value and Item_nth_value in line with Item_lead_lag. This
necessitates the need for defining ::val_json for these functions too.