Bug#26116415 TABLESPACE CLAUSE CONFUSES MYSQLPUMP
Before this fix,
mysqlpump failed to export tables using a tablespace.

The root cause is that mysqlpump parses the
result of SHOW CREATE TABLE,
and is not expecting the TABLESPACE clause.

In the following statement:
CREATE TABLE `t1` (
  `a` int(11) DEFAULT NULL
) /*!50100 TABLESPACE `mytbsp` */ ENGINE=InnoDB DEFAULT CHARSET=latin1

MYSQLPUMP was looking for ") ENGINE=" ...,
and failed to detect this line.

The fix is to relax parsing of SHOW CREATE TABLE,
so that the line:
  ) /*!50100 TABLESPACE `mytbsp` */ ENGINE=...
is also detected as the table options clause.