WL#8069 More work on ndb_print_backup_file More paths coverage tested Various debug flags reset
We relied on that the GCI received in START_NODE_LCP_REQ could
be used as maxGciCompleted. This wasn't true since we could be
in the final phase of a local LCP as part of the restart at this
point in time. Given that this was merely an optimisation, the
code to check this variable was simply removed to ensure that
even if slightly more GCIs have to be executed during restarts,
the restarts will be much safer.