BUG #26818787: ASSERTION: DATA0DATA.IC:430:TUPLE
Fix for an issue with purge of secondary index with virtual and
externally stored column(s). If new record was being inserted
at place of a deleted one (insert by modify) and purge was trying
to build index entry from that record at the same time a NULL
was returned by row_build_index_entry and then used.