BUG#27030339 BACKUP_LOCK IS ALLOWING REMOVAL OF BINARY AND RELAY LOG              FILES
Problem
-------

According to BACKUP_LOCK specification, when the it is taken, no files
must be created, renamed, or removed.

One exception is that new binary log files can be created, but there are
no exceptions with respect to removal of binary log files.

The current implemented behavior is allowing:
- Expired binary log files to be automatically removed after binary log
  rotation;
- Unused relay log files to be automatically removed;

Analysis
--------

Any exception to binary log files should also be extended to relay log
files. Relay log files can be created fine, so nothing to be fixed in
this respect.

Any automatic purge or binary and relay log files should be avoided
while the server instance is locked for backup, but related functions
are not taking the backup lock in consideration when evaluating if files
will be automatically purged.

Fix
---

- Created a new Moved MDL_lock_visitor class to check if a MDL lock
  is owned by any thread;
- Created is_instance_locked_for_backup() function;
- Added a check to is_instance_locked_for_backup() on automatic
  binary log file purge routine and also on automatic relay log file
  purge routine.