WL#11215 Support for custom processes in ATRT
Custom processes are any kind of process running during all test execution.
These processes are normally servers receiving commands from a test enabling
load to be distributed through those processes.

To specify a custom process the [cluster_deployment] group is used. Example:

  [cluster_deployment.2node]
  proc:executor=host1,host2

The above group, defines for cluster's ".2node" a custom process called
"executor", having two instances, each in host1 and host 2.

Options for the above-defined custom process can be defined using the
process name. Example:

  [cluster_deployment.executor.2node]
  cmd = Executor.sh
  generate-port = true
  cpuset = 1

  [cluster_deployment.executor.1.2node]
  args = --arg1

  [cluster_deployment.executor.2.2node]
  args = --arg2

The first group defines options that are common to all "executor" processes,
"cmd" specifying the binary to be executed, "args" (not used here) specifying
binary arguments, "generate-port" flag defining if a "--port=N" argument should
be generated, and "cpuset" cpu binding for the process (feature previously
implemented, that remain as was).

The second and third groups, define options that are specific to each
"executor" instance. Any option defined in such groups will overwrite any
previously defined option.

For all defined custom processes, a global variable will be set in the
environment of API and Client processes. This is necessary to allow tests to be
able to discover/connect to those custom processes. Example:

  EXECUTOR_CONNECTSTRING=host1:14100,host2:141001

in which the custom process name is capitalized and the _CONNECTSTRING suffix
added. Each defined host for that process will be added (hosts separated by
comma) and, if "generate-port" option is set to true for each process, the port
number will be set as ":PORT_NO".