Bug#27010574  WINDOW FUNCTIONS: BUG WITH LAST_VALUE AND FRAME
We originally had the restriction that RANGE frames were only allowed
with a single ORDER BY expression.

We lifted that if there is no value involved, but only UNBOUNDED or
CURRENT ROW. This requires an extension of the logic the detect peers,
which is missing: Currently we set up a comparator for the (first)
order by expression and test whether a row is before or after the
frame by an Item_func_{gt, le} depending on ORDER BY's directions
(ASC, DESC).

The patch generalizes the logic to in Window::setup_range_expressions
and Window::before_or_after_frame to multiple ORDER BY expressions

Added repro test, plus some to exercise NULL handling.