Bug#27041288: ASSERTION FAILURE:               ROW0SEL.CC:3955:PREBUILT->MYSQL_PREFIX_LEN <= RECORD_BUFFER..
If a table contains a column whose length is zero, the optimizer may
allocate a record buffer that is too small to hold the columns read by
the query. This leads to an assert failure in debug builds.

The optimizer looks for the last column accessed by the query in order
to find out how much space it needs for each record. To find the last
column, it compares the pointers to the beginning of each of the
columns. When one of the columns has zero length, two columns may have
the same start position. Since the algorithm doesn't have any
tie-breaker for this case, it may end up returning the second to last
column instead of the last column.

The fix is to compare the end of the column instead of the beginning
of the column, since it is the end that is interesting when
calculating the minimum space needed for a record.

Change-Id: Iadc3a0782dd3bcc553993eeb4ecc4b42b0fc07fa