Bug #26399073: MYSQL DOES NOT COMPILE WITH CLANG ON WINDOWS [noclose]
Modify convenience library linking so that one can use lld-link for linking
executables and use MSVC's linker for linking convenience libraries. This is
currently the best of both worlds, as lld-link can't link convenience libraries,
and MSVC's linker has a hard-coded symbol/file limit that chokes on Clang's
output in debug mode.

In sum, this means that it's possible to compile and link MySQL with Clang on
Windows also in debug mode.

Change-Id: If62cadff0db8f3ce80baf18043c4732458319d24