Bug#27010089: ASSERTION FAILURE:               ROW0SEL.CC:2846:TEMPL->MYSQL_COL_LEN == LEN
When evaluating an end-range condition in a scan of a non-covering
secondary index, and the end-range condition references an indexed
virtual column, InnoDB may read the wrong column from the index. This
could lead to assertion failures or wrong results.

This happens because row_search_end_range_check() calls
row_sel_store_mysql_field() with false as the last argument. The type
of that argument has changed from bool to ulint to represent a column
number. The argument value false is implicitly converted to zero, and
is therefore interpreted as the column number zero, which is not
necessarily correct.

The fix is to pass ULINT_UNDEFINED instead of false as the last
argument torow_sel_store_mysql_field(). This follows the pattern from
Bug#25175249, which changed the type of the argument from bool to
ulint, and which also changed callers that passed false as the last
argument to pass ULINT_UNDEFINED instead.

Change-Id: I73f1ab921235e507421f0df807fdc0e0d1614ffe