BUG#26974113 - CONTRIBUTION BY FACEBOOK; READ COMPRESSED PACKETS OF SIZE                0XFFFFFF.
mysqld server when returning a resultset consisting of logical packets
that are multiples of 0xFFFFFF drops a connection under compress mode.
When server/client processes a compressed packet that consists of logical
packets which are multiples of 0xFFFFFF, the last packet in the series is
terminated by a zero length packet. During read of last zero length packet, we
save the byte of the header of zero length and skip to the next header of the
packet. During the next iteration if there are remaining data in buffer to be
processed, the header of the next packet is overwritten with the byte saved
from the header of the zero length packet. This result in the connection
being dropped.

The fix is to restore the byte that belongs to the header of the next packet.
This patch is based on contribution from Facebook Inc.