Bug#26199978 - WRONG ERROR MESSAGE FOR PARTIAL BACKUPS WITH                GTID_MODE = ON
DESCRIPTION
===========
Attempt to take partial backup through mysqlpump on a GTID
enabled server fails reporting,

 "mysqlpump: [ERROR] (1) A partial dump from a server that
  has GTIDs is not allowed.
  Dump process encountered error and will not continue."

If one trusts this message, he would think that there is no
way ("at all") mysqlpump can be used to take partial backup
with GTID enabled server, which is wrong.

ANALYSIS
========
We've two issues here:

1. The error message, as stated above.
2. One(undocumented) way to still do so is to pass
--set-gtid-purged=OFF. Ideally both the values [ON|OFF]
should be allowed even in case of partial backup, depending
upon the use case (the use cases should be part of error
message)

FIX
===
Sql_formatter::format_dump_start() checks for the
gtid-purged mode, so we now error out "only" when this mode
is set to AUTO. The message is as follows.

 "A partial dump from a server that is using GTID-based
 replication requires the --set-gtid-purged=[ON|OFF] option
 to be specified. Use ON if the intention is to deploy a
 new replication slave using only some of the data from the
 dumped server. Use OFF if the intention is to repair a
 table by copying it within a topology, and use OFF if the
 intention is to copy a table between replication topologies
 that are disjoint and will remain so."

The above message is self-explanatory.

NOTE FOR THE DOCUMENTATION TEAM
===============================
Please add the above message for option
"--set-gtid-purged=value" under "mysqlpump Option
Descriptions" section in the documentation pages.