Bug#26919289 Addendum patch
Previous patch for the 'MYSQLD CRASH DURING 'JOIN PUSHDOWN' ANALYSIS'-bug
assured that the m_remap[]-array was initialized upto the 'count' tables
in the query plan.

However, there were other code paths which tried to access m_remap[]
beyond the end of 'known' tables. Typically such cases are found
where the query planner decide to create a (temporary-) materialized subquery
table which is placed after the 'normal' tables in the query plan.

This patch ensures that the illegal table enum value 'MAX_TABLES' is
returned as the table number when accessing such a table.