Bug #27022925 SPJ: 'MATERIALIZED' SEMI JOINS ARE NOT CONSIDDERED FOR JOIN PUSHOWN
The 'Materialized' part of a semi join execution was represented in such a way
that it was not even 'visible' for the code evaluating the pushability
if a query plan.

The pitfall turned out to be that there are two slightly different
'counts' representing the number of tables in a query plan,
'primary_tables' and 'tables'. It turns out that the 'primary_tables'
did not include the materialized tables in its count.

This patch change the init of table 'count' in the class Join_plan
to use 'tables' instead of 'primary_tables'. The materialized
part of the query plan will then become visible.

As a side effect some 'non-real' (temporary) tables added by the
optimizer also becomes visible. Thus the patch also added check
for 'table() != NULL' a couple of places in order to ignore
such non-real tables.