Bug#26931470: PLEASE BACKPORT TO 5.7: DEFAULT LOG-TC-SIZE TOO SMALL ON POWER - 3 * ( 64K PAGE
This is a backport to 5.7 of

Bug#23014086: DEFAULT LOG-TC-SIZE TOO SMALL ON
              POWER - 3 * ( 64K PAGE SIZE) MINIMIUM NEEDED

The problem was that the memory page size for the memory mapped
transaction coordinator was hardcoded to 8k. This did not work
properly on platforms like ARM64 and PowerPC where the page
size is much larger.

This patch fixes the problem by using my_getpagesize() to get
the page size of the current platform rather than using a hardcoded
value.

This has consequences for and changes the behavior of the
--log-tc-size startup option. The minimum and default value
of the option is now 6 * pagesize. This means it is still
24k on Linux (4k pagesize). The value also now has to be
multiples of pagesize rather than multiples of 8k.

Patch based on contribution from Alexey Kopytov.

(cherry picked from commit e3b09cae22163edfc1951641767b776c8545ad49)

Original backport provided by Daniel Black.