Bug #22486025	INNODB: FAILING ASSERTION: KEY_LEN != 0 || FIND_FLAG != HA_READ_KEY_EXACT
This assertion basically comes because innodb doesn't support
index on zero length column.This problem comes only for
innodb intrinsic temporary table. It was decided that we will
presently return an error for the queries which build a
index on a zero length column . Supporting the index on
zero length column will be taken as a separate bug.