Bug#27068017 COLUMN METADATA REGRESSION SEEN IN TEST X.NOTICES_ENABLE
Fix: just record the new, better result.

Cause: in the output of the X plugin,
flags=128 means MYSQLX_COLUMN_FLAGS_MULTIPLE_KEY i.e. column is part
of an index (see  xpl::Streaming_command_delegate::field_metadata()).
In the query:
 select grp,group_concat(a order by a,d-ascii(c)-a) from test.t1
 group by grp order by grp;
"grp" which is received by the client is not directly t1.grp;
it is an intermediate column created by setup_copy_fields() (for GROUP BY), so it is
not indexed (whereas t1.grp is): there's no indexing which the user
may use on this column. To create this column, setup_copy_fields() calls
Field::new_field() which precisely takes this flag down (as well as
some other flags which similarly don't "survive through a tmp
table step", and unlike flags which do survive e.g. UNSIGNED).
So it is proper to not show this flag to the user, for this query.
This is a good side-effect of the fix for Bug#26188578 ; it's because
in that fix, we let item->field point to the new field, not the original
field, in setup_copy_fields().