Bug #25694140  THE IMPLEMENTATION OF QUERYING P_S.REPLICATION_APPLIER_FILTERS IS SUBOPTIMAL
When querying performance_schema.replication_applier_filters or
performance_schema.replication_global_applier_filters tables,
it generates a view on the filters for every row that is
created. This is suboptimal.

We implement the best approach to make sure that the P_S view over
the filters is generated only when the filters are changed. To
per-channel replication filters, actually increasing counter does
not change the structure of the P_S view. So we should hold the
write lock of rpl_channel_filters to generate the P_S view only on
startup, on CHANGE REPLICATION FILTER, on RESET SLAVE ALL, and on
CHANGE MASTER to ..., then we just need to hold a read lock of
rpl_channel_filters to read the P_S view while querying
P_S.replication_applier_filters. To global replication filter,
we should hold the write lock of Rpl_global_filter to generate
its P_S view only on startup and on CHANGE REPLICATION FILTER,
then we just need to hold a read lock of Rpl_global_filter to read
its P_S view while querying P_S.replication_applier_global_filters.
At the same time, we optimize the code to create a derived class
of class Rpl_filter for global replication filter and fix a prone
issue.