Bug #27015730 ASSERT IN LOG0LOG.CC: LOG_SYS->FLUSHED_TO_DISK_LSN >= OLDEST_LSN
During the recovery the write_lsn was set to current LSN
but the flush positions (current_flush_lsn and
flushed_to_disk_lsn) were left unmodified. If nothing was added to
log before the call to log_checkpoint the assertion was raised.

Because the write_lsn was equal to current lsn the log_write_up_to
was doing nothing keeping the flushed_to_disk_lsn untouched. But the
checkpoint code was doing a check that everything up to checkpoint
lsn was flushed.

The fix sets the flushed position after "recovery" to lsn value
so last written adn flushed positions are equal.