Bug#26975848: IFNULL: CRASHES WITH FIRST_VALUE Bug#27062694: WINDOW FUNC: MUCH CRASHING WITH FIRST_VALUE/LAST_VALUE!! Bug#27062796: WINDOW FUNC: CRASH IN STRING::COPY WITH WKB FUNCTIONS
Patch for Bug#26975848 : This same patch fixes the remaining two bugs
An error is generated while evaluating the first argument to Item_func_ifnull.
Without checking for errors, we proceed further to evaluate the second argument
which is a WINDOWING function. And the error is caught here. Windowing
function returns nullptr. But does not set null_value to true. This results
in server exit later.

Solution:
Set null_value to true in ::val_str functions of Item_lead_lag,
Item_first_last_value and Item_nth_value. ::val_str functions are
expected to set null_value for error cases and for an actual null value.
We have an ASSERT in Item::send function to verify the same.
Also check for errors in Item_sum_hybrid::compute and change ::val_* of
Item_sum_hybrid accordingly.