WL#11179: Add support in CPCD for setting cpu affinity
Add support for optionally setting the cpu affinity set of a process via the
'cpuset' argument when defining a process.

'cpuset' accepts a list of CPUs in the format CPU0,CPU1,... and the process
will be locked to those cpus. This feature is supported in Linux only and
setting cpu affinity in other OS will result in a failure to start the process.
If cpuset is not set, the default behavior of allowing the process to run
freely on any cpu applies.
No validation is done for invalid CPU values or format (e.g. -1, a number
higher than the total available CPUs, or any formatting issue).

Additionally fixed the datatype of the 'env' and 'args' command's argument
to make coherent with the types specified in the API.