Bug#22591899: FILE HANDLE LEAK WHEN SETTING SLOW LOG IN MULTIPLE CONNECTIONS.
log open function accessed the log-name system variables without
holding a lock on sysvars. This could lead to a race condition
between
a) several threads trying to change the same logging sys-var
b) several threads trying to change inter-related logging sys-vars
   (e.g. slow log file name, and slow log on/off state)
c) a thread changing a logging sys-var, and another thread running
   FLUSH [ERROR] LOGS (which also read the log name sys vars while
   reopening the log; and without holding a lock on the sys vars)
While this case seems somewhat artificial in that it requires SUPER
and a lot of pointless waffling with the log settings, it's only
proper that we guard against it all the same. New code factors out
the part that reads the sys-var, so it is not run during FLUSH,
and makes sure that that part of the code is run from sys-var setters
while the sys var lock is still held.