Bug #27041382: ASSERTION `!TABLE->IN_USE->IS_ERROR()'                FAILED.
Issue:
------
When a trigger contains incorrect syntax and the
subsequent INSERT doesn't have any column-list specified,
inserting a new row shouldn't be attempted. But this
happens and it hits an assert.

Solution:
---------
check_for_broken_triggers() raises this problem but the
return code is ignored by
prepare_triggers_for_insert_stmt(), since this function
only returns a void.

Fix will be to move Table_trigger_dispatcher::mark_fields()
 (which eventually calls check_for_broken_triggers()) into
the Sql_cmd_insert_base::prepare_inner() similar to the way
it is done in UPDATE. This way, we won't be executing for a
long time with an unhandled exception.