BUG#27091468: CONCURRENCY ISSUE IN GR_FORCE_MEMBER_EXPEL MAKES IT FAIL SPORADICALLY
When there is an error in a server, it is moved to an error state and then
set to read only mode.

The test case expects that a transaction will reach the server after it
has been moved to read only mode. However, it reaches the server before
it is moved to read only mode sometimes thus generating unexpected messages
in the error log and making the test fail.

The test was change to confirm that super_read_only is set before execute
transaction.