Bug#26199978 - WRONG ERROR MESSAGE FOR PARTIAL BACKUPS WITH                GTID_MODE = ON
POST-PUSH FIX
=============
There is a leak in mysqlpump (Bug#23102944) which can be
seen after the tests were pushed in patch to this bug. This
known issue will be worked upon later. Meanwhile to make
the "LeakSanitizer" silent till this leak is fixed, we're
moving the pushed tests from "main.mysqlpump_extended.test"
to a new file "mysqlpump_partial_bkp.test" and making its
entry in "mysql-test/collections/disabled-asan.list"
(applicable only in 8.0 and trunk). We've also included
"not_valgrind.inc" in the header of this new test file for
the same reason.