BUG#27105803: NO CHECK FOR DUPLICATE @@SERVER_UUID WHEN JOINING GROUP
On Group Replication, like on asynchronous replication directly
connected server, all servers that do belong a group must have
unique server_uuid.

The above condition was required but not enforced, this patch closes
that gap. If a server tries to join a group with a server_uuid
already in use by another server, the join does fail.